deploy/requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
PbPython==0.1.9
python-forecastio==1.3.4

